move resources to reports panel
