+++ app/Filament/Reports/Resources/ReportResource/Pages/CreateReport.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Filament\Fodig\Resources\ReportResource\Pages;
+namespace App\Filament\Reports\Resources\ReportResource\Pages;
 
-use App\Filament\Fodig\Resources\ReportResource;
+use App\Filament\Reports\Resources\ReportResource;
 use Filament\Actions;
 use Filament\Resources\Pages\CreateRecord;
