+++ app/Filament/Reports/Resources/ReportResource/Pages/EditReport.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace App\Filament\Fodig\Resources\ReportEntryResource\Pages;
+namespace App\Filament\Reports\Resources\ReportResource\Pages;
 
-use App\Filament\Fodig\Resources\ReportEntryResource;
+use App\Filament\Reports\Resources\ReportResource;
 use Filament\Actions;
 use Filament\Resources\Pages\EditRecord;
 
-class EditReportEntry extends EditRecord
+class EditReport extends EditRecord
 {
-    protected static string $resource = ReportEntryResource::class;
+    protected static string $resource = ReportResource::class;
 
     protected function getHeaderActions(): array
     {
